XWIKI-15912: Improper "rename log" string is displayed when user is performing a page copy
  * Create a common abstract class for Move and Copy job
  * Finish to move deprecated methods from RefactoringScriptService to
legacy
