+++ xwiki-platform-core/xwiki-platform-refactoring/xwiki-platform-refactoring-api/src/main/java/org/xwiki/refactoring/job/AbstractCopyOrMoveRequest.java
@@ -35,6 +35,16 @@ public abstract class AbstractCopyOrMoveRequest extends EntityRequest
     private static final String PROPERTY_DESTINATION = "destination";
 
     /**
+     * @see #isUpdateLinks()
+     */
+    private static final String PROPERTY_UPDATE_LINKS = "updateLinks";
+
+    /**
+     * @see #isUpdateLinksOnFarm()
+     */
+    private static final String PROPERTY_UPDATE_LINKS_ON_FARM = "updateLinksOnFarm";
+
+    /**
      * @return the destination entity, where to move the entities specified by {@link #getEntityReferences()}
      */
     public EntityReference getDestination()
@@ -50,5 +60,49 @@ public abstract class AbstractCopyOrMoveRequest extends EntityRequest
     public void setDestination(EntityReference destination)
     {
         setProperty(PROPERTY_DESTINATION, destination);
+    }
+
+    /**
+     * @return {@code true} if the links that target the old entity reference (before the move) should be updated to
+     *         target the new reference (after the move), {@code false} to preserve the old link target
+     */
+    public boolean isUpdateLinks()
+    {
+        return getProperty(PROPERTY_UPDATE_LINKS, true);
+    }
+
+    /**
+     * Sets whether the links that target the old entity reference (before the move) should be updated to target the new
+     * reference (after the move) or not.
+     *
+     * @param updateLinks {@code true} to update the links, {@code false} to preserve the old link target
+     */
+    public void setUpdateLinks(boolean updateLinks)
+    {
+        setProperty(PROPERTY_UPDATE_LINKS, updateLinks);
+    }
+
+    /**
+     * @return {@code true} if the job should update the links that target the old entity reference (before the move)
+     *         from anywhere on the farm, {@code false} if the job should update only the links from the wiki where the
+     *         entity was located before the move
+     */
+    public boolean isUpdateLinksOnFarm()
+    {
+        return getProperty(PROPERTY_UPDATE_LINKS_ON_FARM, false);
+    }
+
+    /**
+     * Sets whether the job should update the links that target the old entity reference (before the move) from anywhere
+     * on the farm, or only from the wiki where the entity was located before the mode.
+     * <p>
+     * Note that this parameter has no effect if {@link #isUpdateLinks()} is {@code false}.
+     *
+     * @param updateLinksOnFarm {@code true} to update the links from anywhere on the farm, {@code false} to update only
+     *            the links from the wiki where the entity is located
+     */
+    public void setUpdateLinksOnFarm(boolean updateLinksOnFarm)
+    {
+        setProperty(PROPERTY_UPDATE_LINKS_ON_FARM, updateLinksOnFarm);
     }
 }
