+++ xwiki-platform-core/xwiki-platform-refactoring/xwiki-platform-refactoring-api/src/main/java/org/xwiki/refactoring/script/RequestFactory.java
@@ -284,7 +284,6 @@ public class RequestFactory
         return request;
     }
 
-
     /**
      * Creates a request to restore a specified batch of deleted documents from the recycle bin.
      *
